+++ v2/controllers/openingHours.js
@@ -14,6 +14,9 @@ openingHoursRouter.get('/', (request, response) => {
 openingHoursRouter.post('/', (request, response) => {
   let { day, openinghours} = request.body
 
+  if (!request.user) 
+    return response.status(401).end()  
+
   if (!day || !openinghours) {
     return response.status(400).json({ error: 'content missing' })
   }
@@ -30,12 +33,15 @@ openingHoursRouter.post('/', (request, response) => {
     if (err) {
       return response.status(400).json({ error: 'opening hours not created', message: err.message })
     }
-    response.status(201).json({ id: result.insertId, ...newOpeningHours })
+    response.status(201).json({ id: result.insertId, ...newOpeningHours }).end()
   })
 })
 
 openingHoursRouter.put('/:id', (request, response) => {
   let { day, openinghours } = request.body
+
+  if (!request.user) 
+    return response.status(401).end()  
 
   if (!day || !openinghours) {
     return response.status(400).json({ error: 'content missing' })
@@ -54,11 +60,15 @@ openingHoursRouter.put('/:id', (request, response) => {
     if (err) {
       return response.status(400).json({ error: 'opening hours not updated', message: err.message })
     }
-    response.status(201).json({ id: result.insertId, ...openingHours })
+    response.status(201).json({ id: result.insertId, ...openingHours }).end()
   })
 })
 
 openingHoursRouter.delete('/:id', (request, response) => {
+
+  if (!request.user) 
+    return response.status(401).end()  
+  
   openingHoursSql.deleteOpeningHours(request.params.id, (err, result) => {
     if (err) {
       return response.status(400).json({ error: 'opening hours not deleted', message: err.message })
